requirements.txt
streamlit
scikit-learn
pillow
numpy